pyproject.toml
[project]
name = "pipeline-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "huggingface>=0.0.1",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
]
